refactor(Text): remove the interface to index file
