+++ src/components/data/Text/index.tsx
@@ -1,7 +1,29 @@
+import type { ComponentPropsWithRef } from "react";
+
 import { transformedMeasure } from "../../../utilities/transformedMeasure";
+import { Appearance } from "../../../shared/types/Appearance.type";
+
+import { AlignOptions } from "./types/Text.AlignOptions.type";
+import { HtmlElements } from "./types/Text.HtmlElements.type";
+import { TyposOptions } from "./types/Text.TyposOptions.type";
+import { SizesOptions } from "./types/Text.Size.type";
+
 import { StyledText } from "./styles";
 
-import { ITextProps } from "./interfaces/Text.interface";
+export interface ITextProps extends ComponentPropsWithRef<"p"> {
+  children?: React.ReactNode;
+  textAlign?: AlignOptions;
+  margin?: string;
+  padding?: string;
+  as?: HtmlElements;
+  appearance: Appearance;
+  isDisabled?: boolean;
+  type: TyposOptions;
+  size: SizesOptions;
+  cursorHover?: boolean;
+  parentHover?: boolean;
+  ellipsis?: boolean;
+}
 
 const Text = (props: ITextProps) => {
   const {
