--- src/components/data/Text/types/Text.AppearancesOptions.type.ts
@@ -1,3 +1,0 @@
-import { inube } from "@src/shared/tokens";
-
-export type AppearancesOptions = keyof typeof inube.color.text;
